+++ components/app/result/dish-comment-add-section.tsx
@@ -1,10 +1,11 @@
-import { useLocalSearchParams, useRouter } from 'expo-router';
+import { useLocalSearchParams } from 'expo-router';
 import { useEffect, useState } from 'react';
 import { Alert, StyleSheet, TouchableOpacity, View } from 'react-native';
 import { Input } from '@/components/common/input';
 import { Loader } from '@/components/common/loader';
-import { Icon } from '@/components/common/icon';
 import { postNewComment } from '@/api/api';
+import Icon from 'react-native-vector-icons/MaterialIcons';
+import { spanishBlue, spanishBlueAlpha } from '@/constants/colors';
 
 interface DishCommentAddSectionProps {
     refetch: () => void;
@@ -12,8 +13,6 @@ interface DishCommentAddSectionProps {
 
 export function DishCommentAddSection({ refetch }: DishCommentAddSectionProps) {
     const { id } = useLocalSearchParams<{ id: string }>();
-    // const userContext = useUserContext();
-    const router = useRouter();
     const [commentValue, setCommentValue] = useState<string>('');
     const [isPosting, setIsPosting] = useState<boolean>(false);
 
@@ -55,7 +54,9 @@ export function DishCommentAddSection({ refetch }: DishCommentAddSectionProps) {
                         ? <Loader size={30} />
                         : (
                             <TouchableOpacity onPress={onAddComment} disabled={commentValue.length === 0}>
-                                <Icon type="post" size={32} />
+                                <View style={styles.iconContainer}>
+                                    <Icon name="send" color={commentValue.length === 0 ? spanishBlueAlpha : spanishBlue} size={24} />
+                                </View>
                             </TouchableOpacity>
                         )
                     }
@@ -75,5 +76,8 @@ const styles = StyleSheet.create({
     'dish-comment-add-section__send-icon': {
         marginTop: 24,
         marginLeft: 8
+    },
+    iconContainer: {
+        marginTop: 4
     }
 });