+++ components/common/button.tsx
@@ -1,10 +1,12 @@
-import React from 'react';
+import React, { Component, ReactNode } from 'react';
 import { Text, TouchableOpacity, StyleSheet, View, DimensionValue, Alert } from 'react-native';
 import { constantStyles } from '@/constants/styles';
 import { Href, useRouter } from 'expo-router';
 
 interface ButtonProps {
     label: string;
+    icon?: ReactNode;
+    iconLeft?: boolean;
     disabled?: boolean;
     disabledInfo?: string;
     link?: Href<string | Object>;
@@ -12,7 +14,7 @@ interface ButtonProps {
     onClick?: () => void;
 }
 
-export function Button({ label, link, disabled, disabledInfo, width, onClick }: ButtonProps) {
+export function Button({ label, icon, iconLeft = false, link, disabled, disabledInfo, width, onClick }: ButtonProps) {
     const router = useRouter();
 
     const styles = buttonStyles(width ?? 225, 10);
@@ -30,7 +32,13 @@ export function Button({ label, link, disabled, disabledInfo, width, onClick }: 
     return (
         <View style={styles.buttonContainer}>
             <TouchableOpacity style={disabled ? styles.disabledButton : styles.button} onPress={disabled ? onPressDisabled : onClick ?? onPress}>
-                <Text style={styles.buttonText}>{label}</Text>
+                <View style={styles.buttonBox}>
+                    {iconLeft && <View style={styles.iconContainer}>{icon}</View>}
+                    <Text style={styles.buttonText}>
+                        {iconLeft ? ' ' : ''}{label}
+                    </Text>
+                    {!iconLeft && <View style={styles.iconContainer}>{icon}</View>}
+                </View>
             </TouchableOpacity>
         </View>
     );
@@ -53,8 +61,13 @@ function buttonStyles(width: DimensionValue, margin: DimensionValue) {
             width,
             margin
         },
+        buttonBox: {
+            flexDirection: 'row',
+            justifyContent: 'center',
+            alignItems: 'center'
+        },
         buttonText: {
-            color: '#fff', // Kolor tekstu
+            color: '#fff',
             fontSize: 16,
             textAlign: 'center'
         },
@@ -69,6 +82,10 @@ function buttonStyles(width: DimensionValue, margin: DimensionValue) {
             width,
             margin,
             opacity: 0.618
+        },
+        iconContainer: {
+            marginLeft: 4,
+            marginTop: 2
         }
     });
 }